fix various issues with 304 response generation
Fixes an issue with Safari/WebKit holding connections open when a
304 Not Modified response was generated. The Content-Length,
Content-Type, and other entity headers that are supposed to
be omitted from the response were not being removed properly.

Also, the cache was not sending 304 responses upstream in
response to conditional GET requests with If-None-Match/Etag
values. Not sure how I missed that until now.