basically sane handling of HEAD requests
A HEAD request is never passed through to the backend except
when transitioning with pass!. The cache responds to HEAD
requests without invoking the backend at all when the cached
entry is fresh. When no cache entry exists, or the cached
entry is stale and can be validated, the backend is invoked
with a GET request and the HEAD handling is performed right
before the response is delivered upstream.

This probably needs to be refined a bit but its much better
than the current HEAD handling -- none.