fix meta-store hit, entity-store miss bug
Most entity stores can be purged behind rack-cache (e.g.,
memcached automatically purges based on mem use, the disk store
could be purged with a cron job, etc.). This leads to a situation
where the metastore could point to an entity body that no longer
exists. This would result in a 500 previously but now we detect
it in the metastore and act as if it was a straight miss.

Ideally, we would also purge the metastore entry, since it's no
longer valid and will only cause more false-hits to occur, but
that's something we'll have to tackle later.