Remove configuration language entirely
The event system is gone. This was over-engineering on my part,
plain and simple. As more of RFC 2616 was implemented, the more it
became apparent that a configuration language is simply not
necessary. Setting the appropriate Cache-Control headers at the
origin solves a majority of the problems that Varnish solves with
VCL. The other problems can be implemented with Rack middleware
components placed before or after Rack::Cache.

This is going to let me clean house and remove a ton of extra
complexity needed only to prop up the configuration language.