Disable client reload and revalidate requests by default
This was causing all kinds of confusion since refreshing in
the browser always caused a cache miss. People assumed that
rack-cache wasn't working at all. The allow_reload and
allow_revalidate options now default to false. This breaks
with RFC 2616 but is the desired config in a majority of
gateway cache scenarios.