force BINARY encoding when reading from dalli entity store
Need to pull together a strategy for handling encodings. The binary
file tests are failing for dalli without this change because
cache.get appears is returning a utf-8 string when that's the
external encoding. That actually seems like correct behavior but
the disk store is setup for binary only so I'm inclined to roll with
it as the default behavior for now.