allow IMS/INM requests through to backend on miss, fixes #24
Prior to this change, rack-cache would always strip any
If-None-Match and/or If-Modified-Since headers from the request
before passing along to the backend in an attempt to retrieve a
response to put in the cache (304 responses are not currently
cacheable). This approach falls down when the response includes a
'Cache-Control: private' header (Rails default) because the response
cannot be cached but will also never allow for client initiated
validation.

The downside to this change is that the cache will fill in more
slowly when clients make conditional requests for public resources;
it won't fill in until a non-conditional request is received. The
upside is that validation works for non-cacheable private responses.