Accept a nil value for the entitystore option.
If the entity store is set to nil, entities (the response itself) are not stored. Only metadata are stored, which is enough to determine if a cached response is stale and in this case to validate it.

When the entity store is set to nil, if the cached response is fresh or still valid, rack-cache will return an empty body instead of the original response body. It is the responsability of the calling code to check if the response is cached, either from the status or from the x-rack-cache header, and in this case ignore the response body.

Obviously this is intended only for applications where the body of cached responses is irrelevant, only the fact that the cached response is still valid. The advantage of setting the entity store to nil is that the backend will not quickly fill up with cached entities which are not useful for some apps.