Added a new backend for Entity Store. It is called "Dummy".
This new dummy backend does not actually persist request bodies. It just implements the API but does not cache anything. It is useful for use cases in which the body of responses coming from the cache is not interesting.

The main adventage is that this keeps the disk from filling up periodically with persisted response bodies.

This new Dummy backend can be used by setting the entity store option either to nil or to the "dummy:" URL.