Noop backend now returns saner values from #open and #read.
Noop#open returns an empty array. Noop#read returns an empty string.

This brings Noop in line with other backends, that always return rack-compatible bodies from #open (the empty array returned by Noop#open is suitable to be used as a Rack body, because it responds to #each).

Removed some conditionals that were necessary because Noop#read and #open returned nil.

Added a conditional so that MetaStore#store does not try to read the response body from the cache if the entitystore backend is a Noop. It would be a mistake because this backend always returns an empty body; this would mean that the response body would be replaced with an empty body every time a response was stored, and client code would never receive the response body, not even for requests that miss the cache.