When restoring a response from the cache, default body value is [] instead of nil.
Nil is not a valid Rack response body (it doesn't respond to #each). It makes no sense that
the default body is nil if the body argument is not passed. It makes much more sense
to use a default value of [], which is a valid empty body for a Rack response.