Pass Options To Underlying Storage Driver
While implementing compression for `Redis::Rack::Cache` in
redis-store/redis-rack-cache#30, I noticed that it's not possible to
pass options down to the underlying store, resulting in a
less-than-ideal solution for users of the gem to configure whether they
want responses to be sent over the wire compressed to Redis and
decompressed on the way back. To resolve this, I propose that any
additional options not being used by `Rack::Cache` should be passed into
the underlying storage driver as a hash of options, although this is
backwards-compatible for drivers that don't need any additional options.
It would really help us out a lot if we had something like this in
`Rack::Cache` so custom options that only pertain to the underlying
store (such as connection pool or Redis cluster options) can be passed
in the same Hash as everything else in rack-cache.