Add a proc to allow ignoring parts of the query string in the key
A proc for ignoring parts of query strings when generating a key. This is
useful when you have parameters like `utm` or `trk` that don't affect the
content on the page and are unique per-visitor or campaign. Parameters
like these will be part of the key and cause a lot of churn.